third_party/blink/web_tests/fast/css/fontface-properties-expected.txt

Test getting/setting FontFace properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS ahemFace.family is "Ahem"
PASS ahemFace.style is "italic"
PASS ahemFace.weight is "300"
PASS ahemFace.unicodeRange is "U+0-3FF"
PASS ahemFace.variant is "small-caps"
PASS ahemFace.featureSettings is "\"dlig\""
PASS ahemFace.display is "block"


PASS defaultFace.family is "defaultFace"
PASS defaultFace.style is "normal"
PASS defaultFace.weight is "normal"
PASS defaultFace.stretch is "normal"
PASS defaultFace.unicodeRange is "U+0-10FFFF"
PASS defaultFace.variant is "normal"
PASS defaultFace.featureSettings is "normal"
PASS defaultFace.display is "auto"


PASS constructedFace.family is "constructedFace"
PASS constructedFace.style is "oblique"
PASS constructedFace.weight is "bold"
PASS constructedFace.unicodeRange is "U+100-1FF, U+ABCD"
PASS constructedFace.variant is "small-caps"
PASS constructedFace.featureSettings is "\"liga\" 0"
PASS constructedFace.display is "optional"


PASS modifiedFace.family is "modified"
PASS modifiedFace.style is "italic"
PASS modifiedFace.weight is "900"
PASS modifiedFace.unicodeRange is "U+0-3FF"
PASS modifiedFace.variant is "small-caps"
PASS modifiedFace.featureSettings is "\"dlig\", \"liga\" 0"
PASS modifiedFace.display is "fallback"


PASS face.style = '' threw exception SyntaxError: Failed to set the 'style' property on 'FontFace': Failed to set '' as a property value..
PASS face.weight = 'a' threw exception SyntaxError: Failed to set the 'weight' property on 'FontFace': Failed to set 'a' as a property value..
PASS face.unicodeRange = 'U+' threw exception SyntaxError: Failed to set the 'unicodeRange' property on 'FontFace': Failed to set 'U+' as a property value..
PASS face.variant = '???' threw exception SyntaxError: Failed to set the 'variant' property on 'FontFace': Failed to set '???' as a property value..
PASS face.featureSettings = null threw exception SyntaxError: Failed to set the 'featureSettings' property on 'FontFace': Failed to set 'null' as a property value..
PASS face.display = 123 threw exception SyntaxError: Failed to set the 'display' property on 'FontFace': Failed to set '123' as a property value..
PASS promise1 === promise2 is true
PASS promise1 === promise3 is true
PASS "A network error occurred." is "A network error occurred."
PASS successfullyParsed is true

TEST COMPLETE
